Moved as_logical to meta/
