+++ include/boost/simd/meta/as_logical.hpp
